+++ routes/courseRoutes.js
@@ -10,7 +10,7 @@ const { SectionModel } = require('../models/Sections');
 const { ComponentModel } = require('../models/Components');
 const { ExerciseModel } = require('../models/Exercises');
 const { UserModel } = require('../models/Users');
-const { ContentCreatorModel } = require("../models/ContentCreator");
+const { ContentCreatorModel } = require("../models/ContentCreators");
 const requireLogin = require("../middlewares/requireLogin");
 const { IdentityStore } = require("aws-sdk");
 const mongoose = require('mongoose');
