+++ routes/lectureRoutes.js
@@ -8,7 +8,7 @@ const path = require("path");
 
 // Models
 const { ComponentModel } = require("../models/Components");
-const {ContentCreatorApplication,} = require("../models/ContentCreator");
+const {ContentCreatorApplication,} = require("../models/ContentCreators");
 const requireLogin = require("../middlewares/requireLogin");
 const { SectionModel } = require("../models/Sections");
 const { LectureModel } = require("../models/Lecture");
